backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-openai==0.0.2
chromadb==0.4.18
openai==1.6.1
pypdf==3.17.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
